moved galaxymodel.py to /simulate
