+++ lofasm/simulate/galaxymodel.py
